+++ example/android/app/src/main/kotlin/com/example/ekyc_example/MainActivity.kt
@@ -1,12 +1,5 @@
 package com.example.ekyc_example
 
 import io.flutter.embedding.android.FlutterActivity
-import io.flutter.embedding.engine.FlutterEngine
-import com.example.ekyc.EkycPlugin
 
-class MainActivity : FlutterActivity() {
-    override fun configureFlutterEngine(flutterEngine: FlutterEngine) {
-        super.configureFlutterEngine(flutterEngine)
-        flutterEngine.plugins.add(EkycPlugin())
-    }
-}
+class MainActivity : FlutterActivity()